+++ cmd/push.go
@@ -5,57 +5,61 @@ import (
 	"os"
 	"strings"
 
-	"github.com/hashicorp/go-hclog"
-	"github.com/shipyard-run/shipyard/pkg/clients"
 	"github.com/shipyard-run/shipyard/pkg/config"
 	"github.com/shipyard-run/shipyard/pkg/providers"
+	"github.com/shipyard-run/shipyard/pkg/shipyard"
+	"github.com/shipyard-run/shipyard/pkg/utils"
 	"github.com/spf13/cobra"
 )
 
 var pushCmd = &cobra.Command{
-	Use:                   "push [image] [cluster] [network]",
+	Use:                   "push [image] [cluster]",
 	Short:                 "Push a local Docker image to a cluster",
 	Long:                  `Push a local Docker image to a cluster`,
 	Example:               `yard push nicholasjackson/fake-service:v0.1.3 k3s cloud`,
 	DisableFlagsInUseLine: true,
 	Args:                  cobra.MaximumNArgs(3),
 	Run: func(cmd *cobra.Command, args []string) {
-		// TODO this needs validation
+		l := createLogger()
 
-		image := args[0]
-		cluster := args[1]
-		network := args[2]
-
-		fmt.Printf("Pushing image %s to cluster %s\n\n", image, cluster)
-
-		pc := config.NewK8sCluster(cluster)
-		pc.Networks = []config.NetworkAttachment{config.NetworkAttachment{Name: network}}
-
-		dc, err := clients.NewDocker()
-		if err != nil {
-			fmt.Println("Error pushing image: ", err)
+		if len(args) < 2 {
+			fmt.Println("Push requires two arguments [image] [cluster]")
 			os.Exit(1)
 		}
 
-		dt := clients.NewDockerTasks(dc, hclog.Default())
+		image := args[0]
+		cluster := args[1]
 
-		p := providers.NewK8sCluster(
-			pc,
-			dt,
-			nil,
-			nil,
-			createLogger(),
-		)
+		fmt.Printf("Pushing image %s to cluster %s\n\n", image, cluster)
+
+		// find the cluster in the state
+		sc := config.New()
+		err := sc.FromJSON(utils.StatePath())
+		if err != nil {
+			l.Error("No resources are running, start a stack with 'shipyard run [blueprint]'")
+			return
+		}
+
+		dt, _ := shipyard.GenerateClients(l)
+
+		p, err := sc.FindResource(cluster)
+		if err != nil {
+			fmt.Printf("Cluster %s does not exist", cluster)
+			os.Exit(1)
+		}
+
+		cl := providers.NewK8sCluster(p.(*config.K8sCluster), dt.ContainerTasks, dt.Kubernetes, dt.HTTP, l)
 
 		// get the id of the cluster
-		ids, err := p.Lookup()
+		ids, err := cl.Lookup()
 		if err != nil {
 			fmt.Println("Error getting id for cluster")
 			os.Exit(1)
 		}
 
 		for _, id := range ids {
-			err = p.ImportLocalDockerImages(cluster, id, []config.Image{config.Image{Name: strings.Trim(image, " ")}})
+			l.Info("Pushing to container", "id", id, "image", image)
+			err = cl.ImportLocalDockerImages(p.Info().Name, id, []config.Image{config.Image{Name: strings.Trim(image, " ")}})
 			if err != nil {
 				fmt.Println("Error pushing image: ", err)
 				os.Exit(1)