+++ ChangeLog.md
@@ -4,6 +4,20 @@
 
 ### Helm provider
 * Helm provider now uninstalls the chart when deleting a resource, previousuly it was assumed that a chart and cluster would be deleted together
+* Added `exec` command to allow the creation of a shell or execution of a command in a container or pod
+```
+➜ yard-dev exec k8s_cluster.k3s consul-consul-227vz               
+parameters: []string{"k8s_cluster.k3s", "consul-consul-227vz"} - command: []string{}
+2020-02-19T11:45:28.523Z [DEBUG] Image exists in local cache: image=shipyardrun/tools:latest
+2020-02-19T11:45:28.524Z [INFO]  Creating Container: ref=exec-524329800
+2020-02-19T11:45:28.641Z [DEBUG] Attaching container to network: ref=exec-524329800 network=network.cloud
+/ # ls -las
+total 68
+     4 drwxr-xr-x    1 root     root          4096 Feb 19 11:38 .
+     4 drwxr-xr-x    1 root     root          4096 Feb 19 11:38 ..
+     4 drwxr-xr-x    1 root     root          4096 Sep 13 06:21 bin
+```
+* Added `version` command to return the current application verion
 
 ### Bug fixes
 * Improve test quality